Do no perform MediaWiki logout
It is not necessary to check for Special:UserLogout in casLogin()
since the UserLogoutComplete hook is triggered automatically. Further,
the user is already logged out of MediaWiki by that time, and they
only need to be logged out of CAS.

This change was tested on MW 1.23 and appears to be backwards
compatible.