Run casSetup() only when needed
When phpCAS is needed and $casIsSetUp is false, run casSetup() at that
time rather than aborting. This change makes including casSetup() in
LocalSettings.php no longer necessary. Also removed a few unnecessary
require_once's.

This change was tested on MW 1.23 and appears to be backwards
compatible.

If casSetup() is removed from LocalSettings.php, this commit appears
to be compatible with an unmodded MW 1.27 installation. casSetup()
had to be delayed until just before it was needed because of changes
in MW 1.27 to session handling. With this commit, an "Invalid session
ID" exception is avoided, UNLESS a modified installation (e.g., some
extensions) interacts with the session before the phpCAS client is
initialized.