Prevent phpCAS from creating a new, conflicting session
Elimiates the issue described in commit b0277f6c for MW 1.27, in which
an "Invalid session ID" exception is thrown when attempting to log in.
The phpCAS client constructor calls session_start(), which will
overwrite any existing session data if not protected. This was only
sometimes an issue since existing session data is not always present,
depending on modifications made to the installation (e.g., loading
certain extensions).

This change was tested on MW 1.23 and appears to be backwards
compatible.